refactored server.py to break out into separate tool files
